+++ packages/yt/src/api.rs
@@ -1292,8 +1292,8 @@ pub async fn track_endpoint(
 #[serde(rename_all = "camelCase")]
 pub struct YtSearchQuery {
     query: String,
-    offset: Option<usize>,
-    limit: Option<usize>,
+    offset: Option<u32>,
+    limit: Option<u32>,
 }
 
 #[cfg_attr(
@@ -1305,8 +1305,8 @@ pub struct YtSearchQuery {
         params(
             ("moosicbox-profile" = String, Header, description = "MoosicBox profile"),
             ("query" = String, Query, description = "The search query"),
-            ("offset" = Option<usize>, Query, description = "Page offset"),
-            ("limit" = Option<usize>, Query, description = "Page limit"),
+            ("offset" = Option<u32>, Query, description = "Page offset"),
+            ("limit" = Option<u32>, Query, description = "Page limit"),
         ),
         responses(
             (
